Kepengasuhan

See merge request smart-pesantren/smart-pesantren-be!13

+++ src/main/java/id/smartpesantren/service/SubjectScheduleService.java
@@ -2,21 +2,31 @@ package id.smartpesantren.service;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import id.smartpesantren.constant.LogActivityStatus;
 import id.smartpesantren.dto.ActivityScheduleDTO;
 import id.smartpesantren.entity.*;
 import id.smartpesantren.repository.SubjectScheduleCustomRepository;
+import id.smartpesantren.repository.SubjectScheduleHistoryRepository;
 import id.smartpesantren.repository.SubjectScheduleRepository;
 import id.smartpesantren.security.SecurityUtils;
 import id.smartpesantren.web.rest.vm.SubjectScheduleVM;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class SubjectScheduleService {
     @Autowired
     SubjectScheduleRepository subjectScheduleRepository;
+
+    @Autowired
+    SubjectScheduleHistoryRepository subjectScheduleHistoryRepository;
+
+    @Autowired
+    SubjectScheduleHistoryService subjectScheduleHistoryService;
 
     @Autowired
     private SubjectScheduleCustomRepository customRepository;
@@ -35,9 +45,18 @@ public class SubjectScheduleService {
         }
     }
 
+    @Transactional
     public SubjectScheduleVM saveOrUpdate(SubjectScheduleVM vm) {
         SubjectSchedule ss = fromVm(vm);
         subjectScheduleRepository.save(ss);
+        // Save History
+        SubjectScheduleHistory sh = subjectScheduleHistoryService.fromOrigin(ss);
+        if(vm.getId() == null) {
+            sh.setLogActivity(LogActivityStatus.INSERT);
+        } else {
+            sh.setLogActivity(LogActivityStatus.UPDATE);
+        }
+        subjectScheduleHistoryRepository.save(sh);
         vm.setId(ss.getId());
         return vm;
     }
@@ -57,8 +76,15 @@ public class SubjectScheduleService {
         return ss;
     }
 
+    @Transactional
     public void deleteById(String id) {
-        this.subjectScheduleRepository.deleteById(id);
+        Optional<SubjectSchedule> ss = subjectScheduleRepository.findById(id);
+        if(ss.isPresent()) {
+            SubjectScheduleHistory sh = subjectScheduleHistoryService.fromOrigin(ss.get());
+            sh.setLogActivity(LogActivityStatus.DELETE);
+            this.subjectScheduleRepository.deleteById(id);
+            this.subjectScheduleHistoryRepository.save(sh);
+        }
     }
 
 }
